requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
babel==2.17.0
backoff==2.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==44.0.3
Delorean==1.0.0
docopt==0.6.2
fastapi==0.121.1
ftfy==6.3.1
getmac==0.9.5
google-cloud==0.34.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.14.0
idna==3.11
jproperties==2.1.2
langcodes==3.5.0
language_data==1.3.0
locate==1.1.1
marisa-trie==1.3.1
msgpack==1.1.2
newsapi-python==0.2.7
numpy==2.3.4
pandas==2.3.3
protobuf==6.33.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.5
pytz==2025.2
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
stomp.py==8.2.0
tenacity==9.1.2
tigeropen==3.4.9
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
websocket-client==1.9.0
wordfreq==3.1.1
